+++ src/HeroAPI/HeroAPI/HeroAPI/DataAccessLayer/Models/Power.cs
@@ -3,10 +3,12 @@
 
     public class Power
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string? Name { get; set; }
 
-        public ICollection<HeroPower> HeroPowers { get; set; } = new List<HeroPower>();
+        public ICollection<HeroPower> HeroPowers { get; set; } = new List<HeroPower> ();
+
+        //public ICollection<Hero> Heroes { get; set; } = new ();
 
     }
 }
